+++ Develop/routes/notes.js
@@ -1,0 +1,56 @@
+const notes = require('express').Router();
+const { v4: uuidv4 } = require('uuid');
+const {
+    readFromFile,
+    readAndAppend,
+    writeToFile,
+} = require('../helpers/fsUtils');
+
+// DELETE Route for a specific note
+notes.delete('/:note_id', (req, res) => {
+    const noteId = req.params.note_id;
+    readFromFile('./db/db.json')
+        .then((data) => JSON.parse(data))
+        .then((json) => {
+            // Make a new array of all notes except the one with the ID provided in the URL
+            const result = json.filter((note) => note.note_id !== noteId);
+            // Save that array to the filesystem
+            writeToFile('./db/db.json', result);
+            // Respond to the DELETE request
+            res.json(`Item ${noteId} has been deleted 🗑️`);
+        });
+});
+
+// This API route is a GET Route for retrieving all the notes
+notes.get('/api/notes', (req, res) => {
+    console.info(`${req.method} request received for notes`);
+    readFromFile('./db/db.json').then((data) => res.json(JSON.parse(data)));
+  });
+  
+// This API route is a POST Route for a new UX/UI tip
+notes.post('/api/notes', (req, res) => {
+    console.info(`${req.method} request received to add a note`);
+
+    const { title, text } = req.body;
+
+    if (title && text) {
+        const newNote = {
+            title,
+            text,
+            note_id: uuidv4(),
+        };
+
+        readAndAppend(newNote, './db/db.json');
+        const response = {
+            status: 'success',
+            body: newNote,
+        };
+        res.status(201).json(response);
+    }
+    else {
+        res.json('Error in posting feedback');
+    }
+
+});
+
+module.exports = notes;
